move components into lib
