+++ lib/Cover.tsx
@@ -1,9 +1,9 @@
-import { h, Fragment } from "preact";
+import { h, Fragment } from 'preact'
 
 type Props = {
-	src: string;
-	playing: boolean;
-};
+	src: string
+	playing: boolean
+}
 
 export function Cover({ src, playing }: Props) {
 	return (
@@ -14,10 +14,10 @@ export function Cover({ src, playing }: Props) {
 					width: 100%;
 					overflow: hidden;
 					border-radius: 2px;
-					box-shadow: ${playing ? "0px 1px 4px rgb(0 0 0 / 20%)" : "none"}
+					box-shadow: ${playing ? '0px 1px 4px rgb(0 0 0 / 20%)' : 'none'}
 				}
 			`}</style>
 			<img className="cover" src={src} />
 		</Fragment>
-	);
+	)
 }
