+++ lib/Pause.tsx
