+++ lib/TrackInfo.tsx
